Created python file to interface with databse
The python file will have a class which will be able to interace with
the sqlite3 database. Currently the only member of the class creates
the tables and lays out the schema.

Future Plans:
	The next steps will be to have the class be able to read information
from the tables as well as enter some information. For the time being,
this information may not be formated correctly, its just for general
testing in the other parts of the application.