Added functionality to inserting member functions
From the skeleton code in the previous commit, I was able to add
functionality to the inserting member functions. Now data for users,
rooms, pictures, and devices can be added to the database. As of now
they will all require every bit of data associated with its respective
table. Once the GUI interface is made, it will send data using these
functions.

Some testing member functions were also added for easier development.

One thing to notice, the README.md now has a section at the bottom for
backups. This will not be considered for the final project, but will
require attention at some point.

Goals for next commit:
	The next set from here is adding functionality to the retrieval
functions so that information can be displayed back to the users.