Added preliminary get functions
Two different styles for information retrieval has been added to this
commit. One is to show my initial design while the second is used to
show how I want to implement the rest.

getUsers uses the get all always ideology where as the getEntries has
limitations. This is because this table will get quite large over the
course of the year and in order to perserve performance it limits whats
retrieved from the table. Using a while loop, it check to see if the
user gave some restraint which will be utilized in the GUI

Future commit:
	The goal for the next commit is to have this second implementation
added to the remaining get functions.