Added functionality to rest of database library
After some deliberation, I decided that not every get function was to
make use of the request buffer. For the devices and bedrooms tables, it
felt overkill to implement the buffer as each house will only have a
limited number of each. As for the users and entries, eventually it will
become neccessary.

This file is now ready to be used from other python scripts.