Added basic layout for GUI
The GUI now has the bare basics for how the application will look in the
end. There are two main parts, the menu bar and the working window. The
menu bar allows user to select what part of the unit they wish to see
and the working plan has everything associated with that part.

Using a List Widget, I can create a nice side menu and then place that
inside a HBoxLayout to add whatever I want in the working plane.

Goals for the future:
	As of now the twp layouts are both half and half. Eventually I'll want
the menu to only be about 30% of the left and the rest be the working
plane.
	Additionally, the text in the List widget is small and boring. If I
have time I'd like to add styling to the layout.