Added functionality to the remove user function
The user can now delete users from the database via use of the kill user
button in the table wiget. This will both remove the user from the table
and the database.

Minor changes:
	Appaently there was never a class call to the database class but
rather a complex system of inheritence of the database of a QWidget.
To fix this, an object instance of the database was created and the
database calls were edited to work of the object rather than the
QWidget.

The database interface file also required some edits as not all the
required functions were implemented. For starters there was no functions
to remove users from the database.

The form layout was also changed a little bit to allow for the add user
button to be enabled only after every field is filled. By having each
field have a signal call to the fucntion and then having that function
check for each fields content, the code is now able to detect
completeness.

Future TODO's:
	In the database interface file, the get functions should make use of
the SQL function of limiting results rather than using python loops and
what not.

The database file needs a way to tell calling scripts if there was an
issue with connection. To fix this, there are couple of options. The
error message can either be sent to the calling script where it can then
be brought up as a dialog box/error message.