Modifided imagezmq package to stream videa
The raspberry pi can now live stream videa via tcp and the messenging
queue app zmq to a different machine for analysis.

Thanks to the help of Adrian Rosenbeck and his homemade opencv package:
imutils, video is able to be streamed from the pi throgh a zmq
connection.

Taking only the bits and pieces from imutils which involve operating the
picamera, I was able to frakenstein a script so that only the picamera
module was used and there fore avoided having to reinstall opencv on the
pi zero.

As for the zmq stuff, there was anther module, imagezmq, which was
created specifically for sending images through zmq, that I was also
able to frakenstein.

Now the streamed video can be piped through a facial recognition script
for further analysis.