Added button functionality to the website
The web server can now detect button clicks from the web site and act
accordingly. There was also logging added to the web component.

Currently the flask server only responds with a print statement, but
once the hardware comes in, the script will be able to open doors and
what not.