Finalized the manage rooms tab.
The rooms tab will now display all of the rooms and devices with a
single tree widget rather than the tree table combe. This allows for
better viewing and handeling

The add and remove device pop ups will now also update the table and
database.

Speaking of database, guess I misspelled one of the columns in the
remove device function which caused some issues. It is now spelled
corectly.